pyproject.toml
[tool.poetry]
name = "lm-theory"
version = "0.1.0"
description = ""
authors = ["Simon Ljungbeck <simon.ljungbeck@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain-openai = ">=0.1.0,<0.2.0"
langchain = ">=0.1.13,<0.2.0"
pydantic = "^2.8.2"
regex = "^2024.5.15"
jinja2 = "^3.1.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
